+++ services/ui/frontend/src/components/ServiceGraph/index.js
@@ -1,0 +1,3 @@
+import {GraphPage} from "./Graph";
+
+export {GraphPage};
